Ensure remove am_policy_hub marker file on failed bootstrap

+++ libpromises/bootstrap.c
@@ -83,22 +83,26 @@ bool WriteAmPolicyHubFile(const char *workdir, bool am_policy_hub)
     char *filename = AmPolicyHubFilename(workdir);
     if (am_policy_hub)
     {
-        Log(LOG_LEVEL_INFO, "Assuming role as policy server, with policy distribution point at %s/masterfiles", CFWORKDIR);
-        if (creat(filename, 0600) == -1)
+        if (!GetAmPolicyHub(workdir))
         {
-            Log(LOG_LEVEL_ERR, "Error writing marker file '%s'", filename);
-            free(filename);
-            return false;
+            if (creat(filename, 0600) == -1)
+            {
+                Log(LOG_LEVEL_ERR, "Error writing marker file '%s'", filename);
+                free(filename);
+                return false;
+            }
         }
     }
     else
     {
-        Log(LOG_LEVEL_INFO, "Not assuming role as policy server");
-        if (unlink(filename) != 0)
+        if (GetAmPolicyHub(workdir))
         {
-            Log(LOG_LEVEL_ERR, "Error removing marker file '%s'", filename);
-            free(filename);
-            return false;
+            if (unlink(filename) != 0)
+            {
+                Log(LOG_LEVEL_ERR, "Error removing marker file '%s'", filename);
+                free(filename);
+                return false;
+            }
         }
     }
     free(filename);
